Update 101FailHEPEvt.sh for test harness
Test is expected to pass

+++ validate/fast-tests/101FailHEPEvt.sh
@@ -3,17 +3,36 @@
 # Write a ROOT file in the default geometry.  This output file should
 # be tested to make sure that all of the fields are readable.
 
-OUTPUT=101FailHEPEvt.root
+# Get the directory containing the script from the command line
+# parameters (avoids bash trickery).  Use the current directory as the
+# default.
+DIR=.
+if [ ${#1} -gt 0 ]; then
+    DIR=${1}
+fi
+
+# Make sure that edep-sim has been setup.
+if ! which edep-sim; then
+    echo FAIL: Executable not found for edep-sim
+    exit 1
+fi
+
+# Setup the base names 
+BASE=101FailHEPEvt
+
+OUTPUT=${BASE}.root
 
 if [ -f ${OUTPUT} ]; then
     rm ${OUTPUT}
 fi
 
-cat > 101FailHEPEvt.txt <<EOF
+INPUT=${BASE}.txt
+cat > ${INPUT} <<EOF
 1.0 13.0 0.0 0.0 0.0 0.0 1.0392633396995 0.09623568146043286 0.3492655213021866 1.105658 0.105658 482.8700598805462 331.1324933745929 300.63208929581367 2.9219691707960083
 EOF
 
-cat > 101FailHEPEvt.mac <<EOF
+MACRO=${BASE}.mac
+cat > ${MACRO} <<EOF
 #######################################
 # Set the hit segment.
 #######################################
@@ -21,7 +40,7 @@ cat > 101FailHEPEvt.mac <<EOF
 /edep/hitLength drift 1.0 mm
 /edep/update
 
-/generator/kinematics/hepevt/input 101FailHEPEvt.txt
+/generator/kinematics/hepevt/input ${INPUT}
 /generator/kinematics/hepevt/verbose 2
 /generator/kinematics/set hepevt
 
@@ -30,8 +49,8 @@ cat > 101FailHEPEvt.mac <<EOF
 /generator/add
 EOF
 
-edep-sim -o ${OUTPUT} -C -e 3 101FailHEPEvt.mac | tee 101FailHEPEvt.output 
+edep-sim -o ${OUTPUT} -C -e 3 ${MACRO} | tee ${BASE}.output 
 
-grep "ERROR:.*EDepSimHEPEVT.*Syntax error" 101FailHEPEvt.output || exit 1
+grep "ERROR:.*EDepSimHEPEVT.*Syntax error" ${BASE}.output || exit 1
 
 echo SUCCESS
